ugc2/app/requirements.txt
fastapi==0.115.0
pydantic==2.9.2
uvicorn[standard]==0.31.0
gunicorn==23.0.0
urllib3==1.26.15
pydantic-settings==2.5.2
orjson==3.10.7
typer==0.12.5
requests==2.32.3
PyJWT==2.9.0
email_validator==2.2.0
motor==3.6.0
